% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.nlist}
\alias{aggregate.nlist}
\title{Aggregate nlist}
\usage{
\method{aggregate}{nlist}(x, fun = mean, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{...}{Additional arguments passed to fun.}
}
\value{
An named list of numeric scalars
}
\description{
\lifecycle{defunct}
}
\details{
Replace by \code{\link[=estimates.nlist]{estimates.nlist()}}.
}
\seealso{
Other aggregate:
\code{\link[=aggregate.nlists]{aggregate.nlists()}}
}
\concept{aggregate}
