% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.nlists}
\alias{aggregate.nlists}
\title{Aggregate nlists}
\usage{
\method{aggregate}{nlists}(x, fun = mean, ..., by_chain = FALSE)
}
\arguments{
\item{x}{An object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{...}{Unused.}

\item{by_chain}{A flag specifying whether to aggregate by chains.}
}
\value{
An nlist object if \code{by_chain = FALSE} otherwise an nlists object.
}
\description{
\lifecycle{defunct}
}
\details{
Replace by \code{\link[=estimates.nlists]{estimates.nlists()}}.
}
\seealso{
Other aggregate:
\code{\link[=aggregate.nlist]{aggregate.nlist()}}
}
\concept{aggregate}
