% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{as.data.frame.nlist}
\alias{as.data.frame.nlist}
\title{Coerce nlist Object to Data Frame}
\usage{
\method{as.data.frame}{nlist}(x, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{...}{Unused.}
}
\value{
A data.frame.
}
\description{
Coerces an nlist object to a data.frame with an term column and a value column.
}
\examples{
as.data.frame(nlist(x = 1, y = 4:6))
}
