% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc-list.R
\name{as.mcmc.list.nlist}
\alias{as.mcmc.list.nlist}
\title{As mcmc.list Object}
\usage{
as.mcmc.list.nlist(x, ...)
}
\arguments{
\item{x}{A nlist object.}

\item{...}{Unused.}
}
\value{
An mcmc.list object.
}
\description{
Coerces an nlist object to a \code{coda::mcmc.list} object.
}
\examples{
coda::as.mcmc.list(nlist(x = matrix(1:6, 2)))
}
\seealso{
\code{\link[=nlist-object]{nlist-object()}} and \code{\link[coda:mcmc]{coda::mcmc()}}
}
