% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc-list.R
\name{as.mcmc.list.nlists}
\alias{as.mcmc.list.nlists}
\title{As mcmc Object}
\usage{
\method{as.mcmc}{list.nlists}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}
}
\value{
An mcmc object.
}
\description{
Coerces an nlists object to a \code{coda::mcmc} object.
}
\examples{
coda::as.mcmc.list(nlists(
  nlist(x = matrix(1:6, 2)),
  nlist(x = matrix(3:8, 2))
))
}
\seealso{
\code{\link[=nlists-object]{nlists-object()}} and \code{\link[coda:mcmc]{coda::mcmc()}}
}
