% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-nlists.R
\name{as.nlists}
\alias{as.nlists}
\title{Coerce to nlists}
\usage{
as.nlists(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
An nlist object.
}
\description{
Coerce an R object to an nlists object.
}
