% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term.R
\name{as.term.nlist}
\alias{as.term.nlist}
\title{Coerce to a Term Vector}
\usage{
\method{as.term}{nlist}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\description{
Coerces an R object to a \code{\link[term:term-vector]{term-vector()}}.
}
\section{Methods (by class)}{

\itemize{
\item \code{character}: Coerce character vector to term vector

\item \code{numeric}: Coerce default object to term vector
}
}

\examples{
as.term(nlist(x = matrix(1:4, ncol = 2)))
}
\seealso{
\code{\link[term:term-vector]{term-vector()}} and \code{\link[term:repair_terms]{repair_terms()}}
}
