% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-list.R
\name{as_list.nlists}
\alias{as_list.nlists}
\title{As List}
\usage{
\method{as_list}{nlists}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A list.
}
\description{
Coerces an object to an list.
All attributes are removed except any names.
}
\examples{
as_list(nlists(nlist(x = 1)))
as_list(nlists(nlist(x = 1), nlist(x = 3)))
}
