% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc.R
\name{as_mcmc}
\alias{as_mcmc}
\alias{as_mcmc.nlist}
\alias{as_mcmc.nlists}
\title{Coerce to nlist}
\usage{
as_mcmc(x, ...)

\method{as_mcmc}{nlist}(x, ...)

\method{as_mcmc}{nlists}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}
}
\value{
An nlist object.
}
\description{
Coerce an R object to an mcmc object.
}
\section{Methods (by class)}{
\itemize{
\item \code{nlist}: nlist object to an mcmc object.

\item \code{nlists}: nlists object to an mcmc object.
}}

\examples{
as_nlist(list(x = 1:4))
as_nlist(c(`a[2]` = 3, `a[1]` = 2))
as_mcmc(nlist(x = matrix(1:6, 2)))
as_mcmc(nlists(
  nlist(x = matrix(1:6, 2)),
  nlist(x = matrix(3:8, 2))
))
}
