% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term.R
\name{as_term.nlist}
\alias{as_term.nlist}
\title{Coerce to a Term Vector}
\usage{
\method{as_term}{nlist}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}
}
\value{
A term vector of the terms in \code{x}.
}
\description{
Coerce to a Term Vector
}
\examples{
as_term(nlist(x = matrix(1:4, ncol = 2)))
}
\seealso{
Other coerce term:
\code{\link[=as_term.mcmc]{as_term.mcmc()}},
\code{\link[=as_term.nlists]{as_term.nlists()}},
\code{\link[=as_term_frame]{as_term_frame()}},
\code{\link[=as_term_frame.nlist]{as_term_frame.nlist()}},
\code{\link[=as_term_frame.nlists]{as_term_frame.nlists()}}
}
\concept{coerce term}
