% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term-frame.R
\name{as_term_frame.nlist}
\alias{as_term_frame.nlist}
\title{Coerce nlist Object to Data Frame}
\usage{
\method{as_term_frame}{nlist}(x, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{...}{Unused.}
}
\value{
A data.frame.
}
\description{
Coerces an nlist object to a data.frame with an term column
and a value column.
}
\examples{
as_term_frame(nlist(x = 1, y = 4:6))
}
\seealso{
Other coerce term: 
\code{\link{as_term.mcmc}()},
\code{\link{as_term.nlists}()},
\code{\link{as_term.nlist}()},
\code{\link{as_term_frame.nlists}()},
\code{\link{as_term_frame}()}
}
\concept{coerce term}
