% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term-frame.R
\name{as_term_frame}
\alias{as_term_frame}
\title{Coerce to a Term Frame}
\usage{
as_term_frame(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}
}
\value{
An term_frame object.
}
\description{
A term frame is a tibble with the first column a
term vector called and a numeric column called
value and in the case of an nlists object an integer
vector called samples. It includes the original nlist or nlists object.
}
\seealso{
Other coerce term: 
\code{\link{as_term.mcmc}()},
\code{\link{as_term.nlist}()},
\code{\link{as_term.nlists}()},
\code{\link{as_term_frame.nlist}()},
\code{\link{as_term_frame.nlists}()}
}
\concept{coerce term}
