% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-iterations.R
\name{bind_iterations.mcmc}
\alias{bind_iterations.mcmc}
\title{Bind Iterations}
\usage{
\method{bind_iterations}{mcmc}(x, x2, ...)
}
\arguments{
\item{x}{An object.}

\item{x2}{A second object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The combined object.
}
\description{
Combines two MCMC objects (with the same parameters and chains) by iterations.
}
\examples{
bind_iterations(as_mcmc(nlist(x = 1)), as_mcmc(nlist(x = 3)))
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{estimates}()},
\code{\link[universals]{split_chains}()}
}
