% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_atomic_numeric}
\alias{check_atomic_numeric}
\title{Check Atomic Numeric}
\usage{
check_atomic_numeric(x, length = NA, nas = NA,
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{length}{A flag indicating whether x should have elements 
(versus no elements) or a missing value indicating no requirements.}

\item{nas}{A flag indicating whether x should only have missing values
(versus no missing values) or a missing value indicating no requirements.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Atomic Numeric
}
\examples{
check_atomic_numeric(1)
}
