% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_natomic}
\alias{check_natomic}
\title{Check Atomic Numeric}
\usage{
check_natomic(x, length = NA, nas = NA, x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{length}{A flag indicating whether x should have elements 
(versus no elements) or a missing value indicating no requirements.}

\item{nas}{A flag indicating whether x should only have missing values
(versus no missing values) or a missing value indicating no requirements.}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check Atomic Numeric
}
\examples{
check_natomic(1)
}
