% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_nlist}
\alias{check_nlist}
\title{Check nlist}
\usage{
check_nlist(x, length = NA, sorted = FALSE, class = FALSE,
  x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{length}{A flag specifying whether x should have elements 
(versus no elements) or a missing value indicating no requirements or 
a count or count range of the number of elements or 
a count vector of the permitted number of elements.}

\item{sorted}{xx}

\item{class}{xx}

\item{x_name}{xx}

\item{error}{xx}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if an object is a numeric list.
}
