% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_natomic}
\alias{chk_natomic}
\title{Check Numeric Atomic}
\usage{
chk_natomic(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error 
unless err = FALSE in which case it returns FALSE.
}
\description{
Checks if numeric atomic object.
}
\details{
Checks if numeric object using:

\code{is.numeric(x) && is.atomic(x)}
}
\examples{
chk_natomic(1)
chk_natomic(matrix(1L))
chk_natomic(TRUE, err = FALSE)
}
