% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_nlists}
\alias{chk_nlists}
\title{Check nlists}
\usage{
chk_nlists(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error 
unless err = FALSE in which case it returns FALSE.
}
\description{
Checks if an \code{\link{nlists-object}} while ignoring the class.
}
\examples{
chk_nlists(nlists(nlist(x = 1)))
}
