% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains.mcmc.list}
\alias{collapse_chains.mcmc.list}
\title{Collapse Chains}
\usage{
\method{collapse_chains}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object with one chain.
}
\description{
Collapses an MCMC object's chains into a single chain.
}
\examples{
mcmc_list <- as_mcmc_list(split_chains(nlists(nlist(x = 1), nlist(x = 2))))
collapse_chains(mcmc_list)
}
\seealso{
Other collapse:
\code{\link[=collapse_chains.mcmc]{collapse_chains.mcmc()}},
\code{\link[=collapse_chains.nlist]{collapse_chains.nlist()}},
\code{\link[=collapse_chains.nlists]{collapse_chains.nlists()}}
}
\concept{collapse}
