% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains.nlist}
\alias{collapse_chains.nlist}
\title{Collapse Chains}
\usage{
\method{collapse_chains}{nlist}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object with one chain.
}
\description{
Collapses an MCMC object's chains into a single chain.
}
\details{
As nlist objects can only have 1 chain the object is unchanged.
}
\examples{
collapse_chains(nlist(x = 2))
}
\seealso{
Other collapse: 
\code{\link{collapse_chains.mcmc}()},
\code{\link{collapse_chains.nlists}()}
}
\concept{collapse}
