% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains.nlists}
\alias{collapse_chains.nlists}
\title{Collapse Chains}
\usage{
\method{collapse_chains}{nlists}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object with one chain.
}
\description{
Collapses an MCMC object's chains into a single chain.
}
\examples{
collapse_chains(nlist(x = 2))
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{estimates}()},
\code{\link[universals]{split_chains}()}
}
