% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains}
\alias{collapse_chains}
\alias{collapse_chains.nlist}
\alias{collapse_chains.nlists}
\title{Collapse Chains}
\usage{
collapse_chains(x, ...)

\method{collapse_chains}{nlist}(x, ...)

\method{collapse_chains}{nlists}(x, ...)
}
\arguments{
\item{x}{An MCMC object.}

\item{...}{Unused.}
}
\description{
Collapses an MCMC object's chains into a single chain.
}
\section{Methods (by class)}{
\itemize{
\item \code{nlist}: Collapse chains for an nlist object

As nlist objects can only have 1 chain the object is unchanged.

\item \code{nlists}: Collapse chains for an nlists object
}}

