% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{is.natomic}
\title{Deprecated Functions}
\usage{
is.natomic(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Deprecated Functions
}
\section{Functions}{
\itemize{
\item \code{is.natomic}: Is natomic

\lifecycle{soft-deprecated}

Replace by \code{\link[=is_natomic]{is_natomic()}}
}}

\keyword{internal}
