% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.nlists}
\alias{estimates.nlists}
\title{Estimates}
\usage{
\method{estimates}{nlists}(x, fun = median, ...)
}
\arguments{
\item{x}{An object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{...}{Additional arguments passed to fun.}
}
\value{
An (unsorted) named (by parameters) list of numeric atomic objects of the estimates or an (unsorted) \code{\link[nlist]{nlist}}.
}
\description{
Calculates the estimates for an MCMC object.
}
\examples{
estimates(nlists(nlist(x = 1:3), nlist(x = 2:4)), fun = mean)
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{fill_na}()},
\code{\link[universals]{split_chains}()}
}
