% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.natomic}
\alias{is.natomic}
\title{Is natomic}
\usage{
is.natomic(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether x is a natomic object.
}
\description{
Test whether x is a numeric, atomic object.
An natomic object is a vector, matrix or array of typeof integer or double.
}
\examples{
is.natomic(list(x = 1))
is.natomic(1)
}
