% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.nlists}
\alias{is.nlists}
\title{Is nlists}
\usage{
is.nlists(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether x inherits from nlists.
}
\description{
Test whether x is an \code{\link{nlists_object}}.
}
\examples{
is.nlists(nlist(x = 1))
is.nlists(nlists(nlist(x = 2), nlist(x = 3.5)))
}
