% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains.nlist}
\alias{nchains.nlist}
\title{Number of MCMC Chains of a nlist Object}
\usage{
\method{nchains}{nlist}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of chains of an nlist object.
}
\details{
Always 1L.
}
\examples{
nchains(nlist(x = 1:2))
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{split_chains}()}

Other {terms}: 
\code{\link[universals]{complete_terms}()},
\code{\link[universals]{pars_terms}()}
}
