% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains.nlists}
\alias{nchains.nlists}
\title{Number of MCMC Chains of a nlists Object}
\usage{
\method{nchains}{nlists}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of chains of an nlists object.
}
\examples{
nchains(nlists(nlist(x = c(2, 9)), nlist(x = c(1, 7))))
nchains(split_chains(nlists(nlist(x = c(2, 9)), nlist(x = c(1, 7)))))
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()}

Other {terms}: 
\code{\link[universals]{pars_terms}()}
}
