% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_nlist}
\alias{new_nlist}
\title{New nlist Object}
\usage{
new_nlist()
}
\value{
A zero-length named list of class nlist.
}
\description{
Creates a new zero-length nlist object.
Users are expected to create nlist objects from data using the as.nlist functions.
}
\examples{
new_nlist()
}
