% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist-object.R
\name{nlist-object}
\alias{nlist-object}
\alias{nlist_object}
\title{nlist Object}
\description{
An nlist object is an S3 class list of uniquely named
\code{\link[=natomic_object]{natomic_object()}} elements.
}
\details{
nlist objects are the raw data inputs for analytic engines
such as JAGS, STAN and TMB.
}
\examples{
nlist <- nlist(x = 1, zz = matrix(1:9, 3))
str(nlist)
nlist
}
