% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist-package.R
\docType{package}
\name{nlist-package}
\alias{nlist}
\alias{nlist-package}
\title{nlist: Numeric Lists}
\description{
Create and check numeric lists. A numeric list object is a list of 
  uniquely named numeric atomic objects, ie, integer or double vectors, 
  matrices and arrays. A numeric list represents the raw data format for analytic engines 
  such as JAGS, STAN and TMB.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
