% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist.R
\name{nlist}
\alias{nlist}
\title{Create nlist Object}
\usage{
nlist(...)
}
\arguments{
\item{...}{Uniquely named atomic numeric objects.}
}
\value{
An nlist object.
}
\description{
Creates a nlist object from one of more uniquely named atomic numeric arguments.
}
\examples{
nlist()
}
