% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlists-object.R
\name{nlists-object}
\alias{nlists-object}
\alias{nlists_object}
\title{nlists Object}
\description{
An nlists object is a S3 class list of \code{\link[=nlist_object]{nlist_object()}} elements
with the same names, dimensionalities and typeofs.
}
\details{
nlists objects are useful for storing individual realizations
of a simulated data set.
}
\examples{
nlists(nlist(x = 1, zz = matrix(1:9, 3)), nlist(x = 3.5, zz = matrix(2:10, 3)))
}
