% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlists.R
\name{nlists}
\alias{nlists}
\title{Create nlists Object}
\usage{
nlists(...)
}
\arguments{
\item{...}{nlist objects.}
}
\value{
An nlists object.
}
\description{
Creates a nlists object from one of more nlist objects.
}
\examples{
nlists()
nlists(nlist())
nlists(nlist(x = 1))
nlists(nlist(x = 1), nlist(x = -3))
}
