% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims.nlist}
\alias{npdims.nlist}
\title{Number of Parameter Dimensions}
\usage{
\method{npdims}{nlist}(x, ...)
}
\arguments{
\item{x}{The nlist object}

\item{...}{Unused.}
}
\value{
A named list of the number of dimensions of each numeric atomic object.
}
\description{
Gets the number of dimensions of each \code{\link[=natomic_object]{natomic_object()}} in an
\code{\link[=nlist_object]{nlist_object()}}.
}
\examples{
term::npdims(nlist(x = 1:3))
term::npdims(nlist(y = 3, zz = matrix(2:5, 2)))
}
\seealso{
\code{\link[=pdims]{pdims()}}
}
