% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims.nlists}
\alias{npdims.nlists}
\title{Number of Parameter Dimensions}
\usage{
\method{npdims}{nlists}(x, ...)
}
\arguments{
\item{x}{The nlists object}

\item{...}{Unused.}
}
\value{
A named list of the number of dimensions of each numeric atomic object.
}
\description{
Gets the number of dimensions of each \code{\link[=natomic_object]{natomic_object()}} in an
\code{\link[=nlists_object]{nlists_object()}}.
}
\examples{
term::npdims(nlists(nlist(x = 1:3)))
term::npdims(nlists(
  nlist(y = 3, zz = matrix(2:5, 2)),
  nlist(y = 5, zz = matrix(1:4, 2))
))
}
\seealso{
\code{\link[=pdims]{pdims()}}
}
