% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsams.R
\name{nsams}
\alias{nsams}
\title{Number of Sample Values}
\usage{
nsams(x, ...)
}
\arguments{
\item{x}{The R object.}

\item{...}{Unused.}
}
\value{
A count of the total number of sample values.
}
\description{
Number of Sample Values
}
\examples{
nsams(nlists(nlist(x = c(2,9)), nlist(x = c(1,7))))
}
