% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsims.R
\name{nsims.nlist}
\alias{nsims.nlist}
\title{Number of MCMC Simulations}
\usage{
\method{nsims}{nlist}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of MCMC simulations.
}
\description{
Gets the number of simulations (iterations * chains)
of a Monte Carlo Markov Chain (MCMC) object.
}
\details{
Always 1L.
}
\examples{
nsims(nlist(x = 1:2))
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nterms}()},
\code{\link[universals]{split_chains}()}
}
