% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natomic-object.R
\name{numeric-object}
\alias{numeric-object}
\alias{numeric_object}
\title{numeric Object}
\description{
A numeric object is an integer or double vector, matrix or array.
}
\examples{
mat <- matrix(c(TRUE, FALSE, NA, TRUE), 2L)
mat
is_numeric(mat)
mat <- extras::numericize(mat)
mat
is_numeric(mat)
}
