% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise}
\alias{numericise}
\alias{numericize}
\title{Numericise (or Numericize)}
\usage{
numericise(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
The modified object.
}
\description{
If possible recursively coerces an non-numeric R object to an integer while
preserving its dimensions.
}
\details{
Date, POSIXct and hms objects are floored first.
}
\examples{
numericize(TRUE)
}
