% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{nthin}{A positive integer of the thinning rate.}

\item{...}{Unused.}

\item{flesh}{An atomic vector}

\item{skeleton}{An nlist object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}

\item{chains}{An integer vector of chains.}

\item{iters}{An integer vector of iterations.}

\item{pars}{A character vector of parameter names.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{nas}{A flag specifying whether to also fill missing values.}

\item{parameters}{A character vector (or NULL) of the parameters to subset by.}

\item{iterations}{An integer vector (or NULL) of the iterations to subset by.}

\item{simplify}{A flag specifying whether to drop the sd and zscore
columns. \code{simplify = FALSE} is defunct.}

\item{directional_information}{A flag specifying whether the svalue column
should be calculated using \code{\link[extras:directional_information]{extras::directional_information()}} instead of
\code{\link[extras:svalue]{extras::svalue()}}.
The default value will change from \code{FALSE} to \code{TRUE} in a future release;
set the argument explicitly to avoid the deprecation warning.}

\item{silent}{A flag specifying whether to suppress warning messages.}

\item{by_chain}{A flag specifying whether to aggregate by chains.}
}
\description{
Parameter Descriptions
}
\keyword{internal}
