% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.mcmc.list}
\alias{pars.mcmc.list}
\title{Parameter Names}
\usage{
\method{pars}{mcmc.list}(x, scalar = NULL, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\seealso{
Other parameters: 
\code{\link[universals]{npars}()},
\code{\link[universals]{set_pars}()}
}
