% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.nlist}
\alias{pars.nlist}
\title{Parameter Names}
\usage{
\method{pars}{nlist}(x, scalar_only = FALSE, terms = FALSE, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{scalar_only}{A flag specifying whether to only get the names of
parameters with one term.}

\item{terms}{A flag specifying whether to return the parameter name
for each term.}

\item{...}{Unused.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the parameter names for an \code{\link{nlist_object}}.
}
\examples{
term::pars(nlist(zz = 1, y = 3:6))
}
