% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.nlists}
\alias{pars.nlists}
\title{Parameter Names}
\usage{
\method{pars}{nlists}(x, scalar = NA, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NA), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the parameter names.
}
\examples{
pars(nlists(nlist(zz = 1, y = 3:6), nlist(zz = 4, y = 13:16)))
}
\seealso{
Other {parameters}: 
\code{\link[universals]{npars_scalar}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{pars_scalar}()},
\code{\link[universals]{set_pars}()}
}
