% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims.nlist}
\alias{pdims.nlist}
\title{Parameter Dimensions of an nlist Object}
\usage{
\method{pdims}{nlist}(x, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{...}{Unused}
}
\value{
A named list of the dimensions of each \code{\link[=natomic_object]{natomic_object()}}.
}
\description{
Gets the dimensions of an nlist object.
}
\examples{
term::pdims(nlist(x = 1:3))
term::pdims(nlist(y = 3, zz = matrix(2:5, 2)))
}
\seealso{
\code{\link[term:pdims]{term::pdims()}}
}
