% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims.nlists}
\alias{pdims.nlists}
\title{Dimensions of an nlists Object}
\usage{
\method{pdims}{nlists}(x, ...)
}
\arguments{
\item{x}{An nlists object.}

\item{...}{Unused}
}
\value{
A named list of the dimensions of each \code{\link[=natomic_object]{natomic_object()}}.
}
\description{
Gets the dimensions of an nlists object.
}
\examples{
term::pdims(nlists(nlist(x = 1:3)))
term::pdims(nlists(
  nlist(y = 3, zz = matrix(2:5, 2)),
  nlist(y = 5, zz = matrix(1:4, 2))
))
}
\seealso{
\code{\link[term:pdims]{term::pdims()}}
}
