% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R, R/as-mcmc-list.R, R/as-mcmc.R,
%   R/as-term.R, R/bind-iterations.R, R/collapse-chains.R, R/complete-terms.R,
%   R/estimates.R, R/export.R, R/fill-all.R, R/fill-na.R, R/nchains.R,
%   R/niters.R, R/npdims.R, R/nsims.R, R/nterms.R, R/pars.R, R/pdims.R,
%   R/set-pars.R, R/split-chains.R, R/thin.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aggregate}
\alias{as.mcmc.list}
\alias{as.mcmc}
\alias{as_term}
\alias{as.term}
\alias{bind_iterations}
\alias{collapse_chains}
\alias{complete_terms}
\alias{estimates}
\alias{median}
\alias{zscore}
\alias{lower}
\alias{upper}
\alias{pvalue}
\alias{svalue}
\alias{fill_all}
\alias{fill_na}
\alias{nchains}
\alias{niters}
\alias{npdims}
\alias{nsims}
\alias{nterms}
\alias{pars}
\alias{pdims}
\alias{set_pars}
\alias{pars<-}
\alias{split_chains}
\alias{thin}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coda}{\code{\link[coda:as.mcmc]{as.mcmc()}}, \code{\link[coda:as.mcmc.list]{as.mcmc.list()}}, \code{\link[coda:thin]{thin()}}}

  \item{extras}{\code{\link[extras:fill_all]{fill_all()}}, \code{\link[extras:fill_na]{fill_na()}}, \code{\link[extras:lower]{lower()}}, \code{\link[extras:pvalue]{pvalue()}}, \code{\link[extras:svalue]{svalue()}}, \code{\link[extras:upper]{upper()}}, \code{\link[extras:zscore]{zscore()}}}

  \item{generics}{\code{\link[generics:tidy]{tidy()}}}

  \item{stats}{\code{\link[stats:aggregate]{aggregate()}}, \code{\link[stats:median]{median()}}}

  \item{term}{\code{\link[term:as_term]{as_term()}}, \code{\link[term:as.term]{as.term()}}, \code{\link[term:complete_terms]{complete_terms()}}}

  \item{universals}{\code{\link[universals:bind_iterations]{bind_iterations()}}, \code{\link[universals:collapse_chains]{collapse_chains()}}, \code{\link[universals:estimates]{estimates()}}, \code{\link[universals:nchains]{nchains()}}, \code{\link[universals:niters]{niters()}}, \code{\link[universals:npdims]{npdims()}}, \code{\link[universals:nsims]{nsims()}}, \code{\link[universals:nterms]{nterms()}}, \code{\link[universals:pars]{pars()}}, \code{\link[universals:pars<-]{pars<-()}}, \code{\link[universals:pdims]{pdims()}}, \code{\link[universals:set_pars]{set_pars()}}, \code{\link[universals:split_chains]{split_chains()}}}
}}

