% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relist-nlist.R
\name{relist_nlist}
\alias{relist_nlist}
\title{Relists an unlist nlist Object}
\usage{
relist_nlist(flesh, skeleton)
}
\arguments{
\item{flesh}{An atomic vector}

\item{skeleton}{An nlist object.}
}
\value{
A numeric vector of the values in x.
}
\description{
Relists an nlist object that has been unlisted to a named numeric vector.
Ensures absent terms are included and preserves integer class.
}
\examples{
relist_nlist(c(`a[2]` = 5), nlist(a = 1:3))
}
\seealso{
\code{\link[=as_nlist.numeric]{as_nlist.numeric()}} and \code{\link[=unlist_nlist]{unlist_nlist()}}
}
