% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort.mcmc.list}
\alias{sort.mcmc.list}
\title{Sort mcmc.list Object}
\usage{
\method{sort}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An mcmc.list object.}

\item{...}{Unused.}
}
\value{
An mcmc.list object with the parameters sorted by name.
}
\description{
Sorts an mcmc.list object by its parameter names.
}
\examples{
sort(as_mcmc_list(nlist(y = 1, a = 2:3)))
}
