% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort.mcmc}
\alias{sort.mcmc}
\title{Sort mcmc Object}
\usage{
\method{sort}{mcmc}(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
An mcmc object with the parameters sorted by name.
}
\description{
Sorts an mcmc object by its parameter names.
}
\examples{
sort(as_mcmc(nlist(y = 1, a = 2:3)))
}
