% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains}
\alias{split_chains}
\alias{split_chains.nlists}
\title{Split Chains}
\usage{
split_chains(x, ...)

\method{split_chains}{nlists}(x, ...)
}
\arguments{
\item{x}{An MCMC object.}

\item{...}{Unused.}
}
\description{
Splits each chain in half to double the number chains and halve
the number of iterations.
}
\section{Methods (by class)}{
\itemize{
\item \code{nlists}: Split chains for an \code{\link[=nlists_object]{nlists_object()}}.
}}

