% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.nlists}
\alias{subset.nlists}
\title{Subset nlist}
\usage{
\method{subset}{nlists}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An nlist object to be subsetted.}

\item{pars}{A character vector of the parameters to include in the subsetted vector.}

\item{...}{Unused.}
}
\value{
The modified nlist object.
}
\description{
Subset nlist
}
\examples{
nlists <- nlists(nlist(a = 1, y = 3, x = 1:4),
                nlist(a = 2, y = 4, x = 4:1),
                nlist(a = 3, y = 6, x = 5:2))
subset(nlists)
subset(nlists, "a")
subset(nlists, c("x", "a"))
}
