% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-df.R
\name{term_df.nlist}
\alias{term_df.nlist}
\title{Term Data Frame}
\usage{
\method{term_df}{nlist}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}
}
\value{
A term data frame.
}
\description{
Creates a \code{term-data-frame} from an \code{nlist-object}.
}
\examples{
term::term_df(nlist(x = 1, y = matrix(1:4, 2)))
}
\seealso{
\code{\link{nlist-object}} and \code{\link[term]{term-data-frame}}
}
