% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-df.R
\name{term_df.nlists}
\alias{term_df.nlists}
\title{Term Data Frame}
\usage{
\method{term_df}{nlists}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}
}
\value{
A term data frame.
}
\description{
Creates a \code{term-data-frame} from an \code{nlists-object}.
}
\examples{
term::term_df(nlists(nlist(x = 1, y = matrix(1:4, 2)),
                     nlist(x = 10, y = matrix(0:3, 2))))
}
\seealso{
\code{\link{nlists-object}} and \code{\link[term]{term-data-frame}}
}
