% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.R
\name{thin.default}
\alias{thin.default}
\title{Thin MCMC Object}
\usage{
\method{thin}{default}(x, nthin = 1L, ...)
}
\arguments{
\item{x}{An object.}

\item{nthin}{A positive integer of the thinning rate.}

\item{...}{Unused.}
}
\value{
The thinned MCMC object.
}
\description{
Thins an MCMC object's iterations.
}
\examples{
thin(nlists(nlist(x = 1), nlist(x = 2), nlist(x = 3), nlist(x = 4)), nthin = 2)
}
