% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.mcmc}
\alias{tidy.mcmc}
\title{Turn an object into a tidy tibble}
\usage{
\method{tidy}{mcmc}(x, simplify = TRUE, directional_information = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{simplify}{A flag specifying whether to drop the sd and zscore
columns. \code{simplify = FALSE} is defunct.}

\item{directional_information}{A flag specifying whether the svalue column
should be calculated using \code{\link[extras:directional_information]{extras::directional_information()}} instead of
\code{\link[extras:svalue]{extras::svalue()}}.
The default value will change from \code{FALSE} to \code{TRUE} in a future release;
set the argument explicitly to avoid the deprecation warning.}

\item{...}{Unused.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about model components.
}
\description{
Turn an object into a tidy tibble
}
\section{Methods}{


\Sexpr[stage=render,results=rd]{generics:::methods_rd("tidy")}

}

