% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist.R
\name{unlist.nlist}
\alias{unlist.nlist}
\title{Flatten nlist Object}
\usage{
\method{unlist}{nlist}(x, recursive = TRUE, use.names = TRUE)
}
\arguments{
\item{x}{An nlist object.}

\item{recursive}{Ignored.}

\item{use.names}{A flag specifying whether to name the vector elements.
If \code{TRUE} the elements are named by their terms.}
}
\value{
An atomic numeric vector of the values in x.
}
\description{
Simplifies an nlist object to an atomic vector.
}
\examples{
unlist(nlist(y = 2, x = matrix(4:7, ncol = 2)))
}
