% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist-nlist.R
\name{unlist_nlist}
\alias{unlist_nlist}
\title{Flatten nlist Object}
\usage{
unlist_nlist(x)
}
\arguments{
\item{x}{An nlist object.}
}
\value{
A named numeric vector of the values in x.
}
\description{
Simplifies an nlist object to an named numeric vector
where the names are the terms.
}
\examples{
unlist_nlist(nlist(y = 2, x = matrix(4:7, ncol = 2)))
}
\seealso{
\code{\link[=as_nlist.numeric]{as_nlist.numeric()}} and \code{\link[=relist_nlist]{relist_nlist()}}
}
