% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_natomic}
\alias{vld_natomic}
\alias{vld_nlist}
\alias{vld_nlists}
\title{Validate Numeric Atomic, nlist Object or nlists Object}
\usage{
vld_natomic(x)

vld_nlist(x)

vld_nlists(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object was validated.
}
\description{
Validate Numeric Atomic, nlist Object or nlists Object
}
\section{Functions}{
\itemize{
\item \code{vld_natomic}: Validate Numeric Atomic

\item \code{vld_nlist}: Validate nlist Object

\item \code{vld_nlists}: Validate nlists Object
}}

\examples{
#' vld_natomic
vld_natomic(1)
vld_natomic(matrix(1L))
try(vld_natomic(TRUE))

# vld_nlist
vld_nlist(nlist(x = 1))
try(vld_nlist(list(x = 1)))

# vld_nlists
vld_nlists(nlists(nlist(x = 1)))
vld_nlists(1)
}
\seealso{
\code{\link[=chk_natomic]{chk_natomic()}}
}
