% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_nlist}
\alias{vld_nlist}
\alias{vld_nlists}
\title{Validate nlist Object or nlists Object}
\usage{
vld_nlist(x)

vld_nlists(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object was validated.
}
\description{
Validate nlist Object or nlists Object
}
\section{Functions}{
\itemize{
\item \code{vld_nlists()}: Validate nlists Object

}}
\examples{

# vld_nlist
vld_nlist(nlist(x = 1))
try(vld_nlist(list(x = 1)))

# vld_nlists
vld_nlists(nlists(nlist(x = 1)))
vld_nlists(1)
}
