% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.nlist}
\alias{aggregate.nlist}
\title{Aggregate nlist}
\usage{
\method{aggregate}{nlist}(x, FUN = mean, ...)
}
\arguments{
\item{x}{An R object.}

\item{FUN}{A function that given a numeric vector returns a numeric scalar.}

\item{...}{Additional arguments passed to FUN.}
}
\value{
An named list of numeric scalars
}
\description{
Aggregates an nlist into a named list of numeric scalars.
}
\examples{
aggregate(nlist(x = 1:9))
aggregate(nlist(y = 3:5, zz = matrix(1:9, 3)), FUN = function(x) x[1])
}
