% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.nlist}
\alias{aggregate.nlist}
\title{Aggregate nlist}
\usage{
\method{aggregate}{nlist}(x, fun = mean, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{...}{Additional arguments passed to fun.}
}
\value{
An named list of numeric scalars
}
\description{
Aggregates an \code{\link[=nlist_object]{nlist_object()}} into a named list of numeric scalars.
}
\examples{
aggregate(nlist(x = 1:9))
aggregate(nlist(y = 3:5, zz = matrix(1:9, 3)), fun = function(x) x[1])
}
\seealso{
Other aggregate: 
\code{\link{aggregate.nlists}()}
}
\concept{aggregate}
