% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.nlists}
\alias{aggregate.nlists}
\title{Aggregate nlists}
\usage{
\method{aggregate}{nlists}(x, FUN = mean, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{FUN}{A function that given a numeric vector returns a numeric scalar.}

\item{...}{Additional arguments passed to FUN.}
}
\value{
An nlist object
}
\description{
Aggregates an \code{\link{nlists_object}} into a \code{\link{nlist_object}}.
}
\examples{
aggregate(nlists(nlist(x = 1:3), nlist(x = 2:4)))
}
