% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-nlist.R
\name{as.nlist}
\alias{as.nlist}
\title{Coerce to nlist}
\usage{
as.nlist(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
An nlist object.
}
\description{
Coerce an R object to an nlist object.
}
