% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-nlist.R, R/as-nlists.R
\name{as.nlist}
\alias{as.nlist}
\alias{as.nlist.list}
\alias{as.nlist.data.frame}
\alias{as.nlists.list}
\alias{as.nlists.nlist}
\title{Coerce to nlist}
\usage{
as.nlist(x, ...)

\method{as.nlist}{list}(x, ...)

\method{as.nlist}{data.frame}(x, ...)

\method{as.nlists}{list}(x, ...)

\method{as.nlists}{nlist}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
An nlist object.
}
\description{
Coerce an R object to an \code{\link{nlist_object}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Coerce list to nlist

\item \code{data.frame}: Coerce data.frame to nlist

\item \code{list}: Coerce list to nlists

\item \code{nlist}: Coerce nlist to nlists
}}

\examples{
as.nlist(list(x = 1:4))
}
