% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-nlists.R
\name{as.nlists}
\alias{as.nlists}
\title{Coerce to nlists}
\usage{
as.nlists(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
An nlists object.
}
\description{
Coerce an R object to an \code{\link{nlists_object}}.
}
\examples{
as.nlists(list(nlist(x = c(1,5)), nlist(x = c(2,3)), nlist(x = c(3,2))))
}
