% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term.R
\name{as_term.mcmc}
\alias{as_term.mcmc}
\title{Coerce to a Term Vector}
\usage{
\method{as_term}{mcmc}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}
}
\description{
Coerce to a Term Vector
}
\examples{
as_term(as_mcmc(nlist(x = matrix(1:4, ncol = 2))))
}
