% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_nlists}
\alias{check_nlists}
\title{Check nlist}
\usage{
check_nlists(x, length = NA, nas = NA, class = TRUE,
  x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{length}{A flag indicating the nlist elements should have elements 
(versus no elements) or a missing value indicating no requirements.}

\item{nas}{A flag indicating whether x should only have missing values
(versus no missing values) or a missing value indicating no requirements.}

\item{class}{A flag indicating whether x should inherit from nlists and the elements from nlist
(versus not inherit) or a missing value indicating no requirements.}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Checks if an object is a numeric list.
}
