% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_nlist}
\alias{chk_nlist}
\alias{chk_nlists}
\title{Check nlist Object or nlists Object}
\usage{
chk_nlist(x, x_name = NULL)

chk_nlists(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
\code{NULL}, invisibly. Called for the side effect of throwing an error
if the condition is not met.
}
\description{
\code{chk_nlist}
checks if an \code{\link[=nlist-object]{nlist-object()}}.
}
\section{Functions}{
\itemize{
\item \code{chk_nlists()}: Check nlists Object

\code{chk_nlists}
checks if an \code{\link[=nlists-object]{nlists-object()}}.

}}
\examples{

# chk_nlist
chk_nlist(nlist(x = 1))
try(chk_nlist(list(x = 1)))

# chk_nlists
chk_nlists(nlists(nlist(x = 1)))
}
