% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{is.natomic}
\alias{is.nlist}
\alias{is.nlists}
\title{Deprecated Functions}
\usage{
is.natomic(x)

is.nlist(x)

is.nlists(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Deprecated Functions
}
\section{Functions}{
\itemize{
\item \code{is.natomic()}: Is natomic

\lifecycle{soft-deprecated}

Replace by \code{\link[=is.numeric]{is.numeric()}}

\item \code{is.nlist()}: Is nlist

\lifecycle{soft-deprecated}

Replace by \code{\link[=is_nlist]{is_nlist()}}

\item \code{is.nlists()}: Is nlists

\lifecycle{soft-deprecated}

Replace by \code{\link[=is_nlists]{is_nlists()}}

}}
\keyword{internal}
