% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na.nlists}
\alias{fill_na.nlists}
\title{Fill Missing Values}
\usage{
\method{fill_na}{nlists}(x, value = 0L, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A scalar of the value to replace missing values with.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Fills an object's missing values.
}
\examples{
fill_na(nlist(x = c(2, NA), y = matrix(c(1:3, NA), nrow = 2)))
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{estimates}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{split_chains}()}
}
