% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na}
\alias{fill_na}
\alias{fill_na.default}
\alias{fill_na.logical}
\alias{fill_na.integer}
\title{Fill Missing Values}
\usage{
fill_na(x, ...)

\method{fill_na}{default}(x, value = 0, ...)

\method{fill_na}{logical}(x, value = FALSE, ...)

\method{fill_na}{integer}(x, value = 0L, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{value}{A scalar of the value to replace missing values with.}
}
\value{
The object with missing values filled.
}
\description{
Fills missing values.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Fill missing values of default object

\item \code{logical}: Fill missing values of logical vector

\item \code{integer}: Fill missing values of integer vector
}}

\examples{
nsams(nlist(x = 2))
nsams(nlist(x = 1:2))
nsams(nlists(nlist(x = c(2, 9)), nlist(x = c(1, 7))))
}
