% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.nlist}
\alias{is.nlist}
\title{Is nlist}
\usage{
is.nlist(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether x inherits from nlist.
}
\description{
Test whether x is an \code{\link{nlist_object}}.
}
\examples{
is.nlist(1)
is.nlist(list(x = 1))
is.nlist(nlist(x = 1))
}
