% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist.R
\name{nlist}
\alias{nlist}
\title{Create nlist Object}
\usage{
nlist(...)
}
\arguments{
\item{...}{Uniquely named atomic numeric objects.}
}
\value{
An nlist object.
}
\description{
Creates a \code{\link[=nlist_object]{nlist_object()}} from one of more uniquely named \code{\link[=natomic_object]{natomic_object()}} arguments.
}
\examples{
nlist()
nlist(x = 1)
nlist(y = 1:4, zz = matrix(1:9, 3))
}
