% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist.R
\name{nlist}
\alias{nlist}
\alias{nlist-object}
\alias{nlist_object}
\title{Create nlist Object}
\usage{
nlist(...)
}
\arguments{
\item{...}{Uniquely named numeric objects.}
}
\value{
An nlist object.
}
\description{
Creates a \code{\link[=nlist_object]{nlist_object()}} from one of more uniquely named numeric arguments.
}
\details{
An nlist object is an S3 class list of uniquely named
numeric elements.

nlist objects are the raw data inputs for analytic engines
such as JAGS, STAN and TMB.
}
\examples{
nlist()
nlist(x = 1)
nlist(y = 1:4, zz = matrix(1:9, 3))
}
