% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlists.R
\name{nlists}
\alias{nlists}
\title{Create nlists Object}
\usage{
nlists(...)
}
\arguments{
\item{...}{nlist objects.}
}
\value{
An nlists object.
}
\description{
Creates an \code{\link[=nlists_object]{nlists_object()}} from one of more \code{\link[=nlist_object]{nlist_object()}}s.
}
\examples{
nlists()
nlists(nlist())
nlists(nlist(x = 1))
nlists(nlist(x = 1), nlist(x = -3))
}
