% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsims.R
\name{nsims}
\alias{nsims}
\title{Number of Simulations}
\usage{
nsims(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count of the number of simulations.
}
\description{
Gets the number of simulations.
}
\examples{
nsims(nlists(nlist(x = 1), nlist(x = 4)))
}
