% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms.nlist}
\alias{nterms.nlist}
\title{Number of Terms}
\usage{
\method{nterms}{nlist}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of terms of an \code{\link[=nlist_object]{nlist_object()}}.
}
\examples{
term::nterms(nlist(x = 2))
term::nterms(nlist(x = NA_real_))
term::nterms(nlist(x = 3, zz = matrix(2:5, 2)))
}
