% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms.nlists}
\alias{nterms.nlists}
\title{Number of Terms}
\usage{
\method{nterms}{nlists}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of terms of an \code{\link[=nlists_object]{nlists_object()}}.
}
\examples{
term::nterms(nlists(nlist(x = 1:3)))
term::nterms(nlists(
  nlist(y = 3, zz = matrix(2:5, 2)),
  nlist(y = 5, zz = matrix(1:4, 2))
))
}
