% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{nthin}{A positive integer of the thinning rate.}

\item{...}{Unused.}

\item{flesh}{An atomic vector}

\item{skeleton}{An nlist object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}

\item{chains}{An integer vector of chains.}

\item{iters}{An integer vector of iterations.}

\item{pars}{A character vector of parameter names.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{nas}{A flag specifying whether to also fill missing values.}
}
\description{
Parameter Descriptions
}
\keyword{internal}
