% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-scalar.R
\name{pars_scalar.nlist}
\alias{pars_scalar.nlist}
\title{Scalar Parameters}
\usage{
\method{pars_scalar}{nlist}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the name of each scalar parameter.
}
\details{
Useful for subsetting MCMC objects.
}
\section{Methods (by class)}{

\itemize{
\item \code{term}: Scalar Parameter Names for term vector
}
}

\examples{
pars_scalar(nlist(zz = 1, y = 3:6))
}
\seealso{
Other {parameters}: 
\code{\link[term]{npars_scalar}()},
\code{\link[term]{pars.term}()},
\code{\link[term]{pars_terms}()}
}
