% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims.nlist}
\alias{pdims.nlist}
\title{Parameter Dimensions}
\usage{
\method{pdims}{nlist}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of integer vectors of the dimensions of each parameter.
}
\description{
Gets the dimensions of each parameter of an object.
}
\examples{
pdims(nlist(x = 1:3))
pdims(nlist(y = 3, zz = matrix(2:5, 2)))
}
\seealso{
Other dimensions: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()}
}
