% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R, R/niters.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nchains}
\alias{niters}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{universals}{\code{\link[universals]{nchains}}, \code{\link[universals]{niters}}}
}}

