% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc.R, R/as-term.R, R/collapse-chains.R,
%   R/fill-all.R, R/fill-na.R, R/nchains.R, R/niters.R, R/npdims.R, R/nsims.R,
%   R/nterms.R, R/pars.R, R/pdims.R, R/split-chains.R, R/thin.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.mcmc}
\alias{as.term}
\alias{collapse_chains}
\alias{fill_all}
\alias{fill_na}
\alias{nchains}
\alias{niters}
\alias{npdims}
\alias{nsims}
\alias{nterms}
\alias{pars}
\alias{pdims}
\alias{split_chains}
\alias{thin}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coda}{\code{\link[coda]{as.mcmc}}, \code{\link[coda]{thin}}}

  \item{term}{\code{\link[term]{as.term}}}

  \item{universals}{\code{\link[universals]{collapse_chains}}, \code{\link[universals]{fill_all}}, \code{\link[universals]{fill_na}}, \code{\link[universals]{nchains}}, \code{\link[universals]{niters}}, \code{\link[universals]{npdims}}, \code{\link[universals]{nsims}}, \code{\link[universals]{nterms}}, \code{\link[universals]{pars}}, \code{\link[universals]{pdims}}, \code{\link[universals]{split_chains}}}
}}

