% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc-list.R, R/as-mcmc.R,
%   R/collapse-chains.R, R/fill-na.R, R/nchains.R, R/niters.R, R/nsims.R,
%   R/split-chains.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.mcmc}
\alias{collapse_chains}
\alias{fill_na}
\alias{nchains}
\alias{niters}
\alias{nsims}
\alias{split_chains}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coda}{\code{\link[coda]{as.mcmc}}, \code{\link[coda]{as.mcmc}}}

  \item{universals}{\code{\link[universals]{collapse_chains}}, \code{\link[universals]{fill_na}}, \code{\link[universals]{nchains}}, \code{\link[universals]{niters}}, \code{\link[universals]{nsims}}, \code{\link[universals]{split_chains}}}
}}

