% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.nlist}
\alias{subset.nlist}
\title{Subset nlist}
\usage{
\method{subset}{nlist}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An nlist object to be subsetted.}

\item{pars}{A character vector of the parameters to include in the subsetted vector.}

\item{...}{Unused.}
}
\value{
The modified nlist object.
}
\description{
Subset nlist
}
\examples{
nlist <- nlist(a = 1, y = 3, x = 1:4)
subset(nlist)
subset(nlist, "a")
subset(nlist, c("x", "a"))
}
