% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.nlist}
\alias{subset.nlist}
\title{Subset nlist Object}
\usage{
\method{subset}{nlist}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An nlist object.}

\item{pars}{A character vector of parameter names.}

\item{...}{Unused.}
}
\value{
An nlist object.
}
\description{
Subsets an nlist object by its parameters.
}
\details{
It can also be used to reorder the parameters.
}
\examples{
nlist <- nlist(a = 1, y = 3, x = 1:4)
subset(nlist)
subset(nlist, "a")
subset(nlist, c("x", "a"))
}
