% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.mcmc}
\alias{tidy.mcmc}
\title{Turn an object into a tidy tibble}
\usage{
\method{tidy}{mcmc}(x, simplify = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{simplify}{A flag specifying whether to drop sd and zscore columns.}

\item{...}{Unused.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about model components.
}
\description{
Turn an object into a tidy tibble
}
\section{Methods}{


\Sexpr[stage=render,results=rd]{generics:::methods_rd("tidy")}

}

